+++ Sources/KeyboardKit/System/SystemKeyboardButtonRowItem.swift
@@ -52,9 +52,7 @@ public struct SystemKeyboardButtonRowItem<Content: View>: View {
             .frame(height: item.size.height - item.insets.top - item.insets.bottom)
             .rowItemWidth(for: item, totalWidth: keyboardWidth, referenceWidth: inputWidth)
             .systemKeyboardButtonStyle(
-                for: item.action,
-                appearance: appearance,
-                isPressed: isPressed)
+                appearance.systemKeyboardButtonStyle(for: item.action, isPressed: isPressed))
             .padding(item.insets)
             .background(Color.clearInteractable)
             .keyboardGestures(